feat: update booking-related DTOs and entities to use decimal for monetary values; add new DTOs for booking and availability

+++ PeerTutoringSystem.Application/DTOs/Booking/BookingSessionDto.cs
@@ -17,7 +17,7 @@ namespace PeerTutoringSystem.Application.DTOs.Booking
         public string? Description { get; set; }
         public BookingStatus Status { get; set; }
         public PaymentStatus PaymentStatus { get; set; }
-        public int OrderCode { get; set; }
+        public long OrderCode { get; set; }
         public DateTime CreatedAt { get; set; }
         public string StudentName { get; set; }
         public string TutorName { get; set; }
